requirements.txt
bootstrap-admin==0.3.6
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.8
django-markdown==0.8.4
django-simplemde==0.0.9
gunicorn==19.4.5
Markdown==2.6.5
markdown2==2.3.0
mdx-video==0.1.8
Pillow==3.3.0
Pygments==2.1
pytz==2015.7
static3==0.6.1
